+++ android/app/src/main/kotlin/io/actingweb/actingweb_firstapp/MainActivity.kt
@@ -1,0 +1,6 @@
+package io.actingweb.actingweb_firstapp
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}