+++ integration_test/app_test.dart
@@ -2,24 +2,30 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:firebase_core/firebase_core.dart';
 
 // The application under test.
-import 'package:first_app/main.dart';
+import 'package:first_app/app.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  FlutterError.onError = (FlutterErrorDetails details) {
+    FlutterError.dumpErrorToConsole(details);
+  };
+  Firebase.initializeApp();
   group('first_app', () {
     // First, define the Finders and use them to locate widgets from the
     // test suite. Note: the Strings provided to the `byValueKey` method must
     // be the same as the Strings we used for the Keys.
-    final loginButtonFinder = find.byKey(Key('LoginPage_LoginButton'));
+    final loginButtonFinder = find.byKey(const Key('LoginPage_LoginButton'));
     final startListeningButtonFinder =
-        find.byKey(Key('LocationPage_StartListeningButton'));
-    final locationTileFinder = find.byKey(Key('LocationPage_LocationTile'));
-    final mapToggleButton = find.byKey(Key('OverlayMap_ToggleButton'));
+        find.byKey(const Key('LocationPage_StartListeningButton'));
+    final locationTileFinder =
+        find.byKey(const Key('LocationPage_LocationTile'));
+    final mapToggleButton = find.byKey(const Key('OverlayMap_ToggleButton'));
     final mapOverlayFinder = find.byType(GoogleMap);
     final openDrawerMenuButton = find.byTooltip("Open navigation menu");
-    final exitButtonFinder = find.byKey(Key('DrawerMenuTile_LogOut'));
+    final exitButtonFinder = find.byKey(const Key('DrawerMenuTile_LogOut'));
 
     testWidgets('app test', (tester) async {
       var app = await getApp(mock: true);