+++ lib/ui/pages/home/drawer.dart
@@ -1,37 +1,40 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:first_app/models/appstate.dart';
-import 'package:first_app/generated/l10n.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:first_app/providers/auth.dart';
 import 'package:first_app/ui/widgets/custom_dialog.dart';
 
 class HomePageDrawer extends StatelessWidget {
+  const HomePageDrawer({Key? key}) : super(key: key);
+
   void _userInfo(BuildContext context) {
     var appState = Provider.of<AppStateModel>(context, listen: false);
     var auth0 = AuthClient(
         authClient: Provider.of<AppStateModel>(context, listen: false)
             .mocks
             .getAppAuth());
-    auth0.getUserInfo(appState.userToken).then((res) {
-      if (res != null) {
-        // The demo.identityserver.io/api/test API doesn't return anything
-        // interesting, so we fake the setting of user info
-        Provider.of<AppStateModel>(context, listen: false).setUserInfo(
-            Map.from({
-          'email': S.of(context).drawerEmail,
-          'name': S.of(context).drawerUser
-        }));
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          content: Text(S.of(context).drawerGetUserInfoResultMsg),
-          duration: const Duration(seconds: 3),
-        ));
-      } else {
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          content: Text(S.of(context).drawerGetUserInfoFailedMsg),
-          duration: const Duration(seconds: 3),
-        ));
+    try {
+      var res = auth0.getUserInfo(appState.userToken);
+      if (res is List || res is Map) {
+        // Do something
       }
-    });
+      // The demo.identityserver.io/api/test API doesn't return anything
+      // interesting, so we fake the setting of user info
+      Provider.of<AppStateModel>(context, listen: false).setUserInfo(Map.from({
+        'email': AppLocalizations.of(context)!.drawerEmail,
+        'name': AppLocalizations.of(context)!.drawerUser
+      }));
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+        content: Text(AppLocalizations.of(context)!.drawerGetUserInfoResultMsg),
+        duration: const Duration(seconds: 3),
+      ));
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+        content: Text(AppLocalizations.of(context)!.drawerGetUserInfoFailedMsg),
+        duration: const Duration(seconds: 3),
+      ));
+    }
   }
 
   @override
@@ -42,49 +45,51 @@ class HomePageDrawer extends StatelessWidget {
         children: <Widget>[
           buildDrawerHeader(context),
           ListTile(
-            key: Key("DrawerMenuTile_RefreshTokens"),
-            title: Text(S.of(context).drawerRefreshTokens),
+            key: const Key("DrawerMenuTile_RefreshTokens"),
+            title: Text(AppLocalizations.of(context)!.drawerRefreshTokens),
             onTap: () {
               appState.refresh();
-              Navigator.of(context).pop();
               ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                content: Text(S.of(context).drawerRefreshTokensResultMsg),
+                content: Text(
+                    AppLocalizations.of(context)!.drawerRefreshTokensResultMsg),
                 duration: const Duration(seconds: 3),
               ));
-            },
-          ),
-          ListTile(
-            key: Key("DrawerMenuTile_GetUserInfo"),
-            title: Text(S.of(context).drawerGetUserInfo),
-            onTap: () {
               Navigator.of(context).pop();
-              _userInfo(context);
             },
           ),
           ListTile(
-            key: Key("DrawerMenuTile_Localisation"),
-            title: Text(S.of(context).drawerLocalisation),
-            subtitle: Text(appState.locale!),
+            key: const Key("DrawerMenuTile_GetUserInfo"),
+            title: Text(AppLocalizations.of(context)!.drawerGetUserInfo),
+            onTap: () {
+              _userInfo(context);
+              Navigator.of(context).pop();
+            },
+          ),
+          ListTile(
+            key: const Key("DrawerMenuTile_Localisation"),
+            title: Text(AppLocalizations.of(context)!.drawerLocalisation),
+            subtitle: Text(appState.localeAbbrev!),
             onTap: () {
               // Here you should have a widget to select among
               // supported locales. This is just a quick and dirty
               // switch
               appState.switchLocale();
-              Navigator.of(context).pop();
               ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                content: Text(S.of(context).drawerLocalisationResultMsg +
-                    appState.locale!),
+                content: Text(
+                    AppLocalizations.of(context)!.drawerLocalisationResultMsg +
+                        appState.localeAbbrev!),
                 duration: const Duration(seconds: 3),
               ));
+              Navigator.of(context).pop();
             },
           ),
           ListTile(
-            key: Key('DrawerMenuTile_LogOut'),
-            leading: new Icon(
+            key: const Key('DrawerMenuTile_LogOut'),
+            leading: const Icon(
               Icons.exit_to_app,
               color: Color(0xe81751ff),
             ),
-            trailing: Text(S.of(context).logoutButton),
+            trailing: Text(AppLocalizations.of(context)!.logoutButton),
             onTap: () {
               Navigator.of(context).pop();
               appState.logOut();
@@ -99,18 +104,18 @@ class HomePageDrawer extends StatelessWidget {
 Widget buildDrawerHeader(BuildContext context) {
   var appState = Provider.of<AppStateModel>(context);
   return UserAccountsDrawerHeader(
-    key: Key("DrawerMenu_Header"),
+    key: const Key("DrawerMenu_Header"),
     accountName: Text(appState.name == null
-        ? S.of(context).drawerHeaderInitialName
+        ? AppLocalizations.of(context)!.drawerHeaderInitialName
         : appState.name!),
     accountEmail: Text(appState.email == null
-        ? S.of(context).drawerHeaderInitialEmail
+        ? AppLocalizations.of(context)!.drawerHeaderInitialEmail
         : appState.email!),
     onDetailsPressed: () => showModalBottomSheet(
       context: context,
       builder: (BuildContext context) {
         var container = Container(
-          key: Key("DrawerMenu_BottomSheet"),
+          key: const Key("DrawerMenu_BottomSheet"),
           alignment: Alignment.center,
           height: 200,
           decoration: BoxDecoration(
@@ -120,56 +125,63 @@ Widget buildDrawerHeader(BuildContext context) {
             children: <Widget>[
               ListTile(
                 title: Text(appState.name == null
-                    ? S.of(context).drawerEmptyName
+                    ? AppLocalizations.of(context)!.drawerEmptyName
                     : appState.name!),
                 subtitle: Text(appState.email == null
-                    ? S.of(context).drawerEmptyEmail
+                    ? AppLocalizations.of(context)!.drawerEmptyEmail
                     : appState.email!),
               ),
               ListTile(
-                title: Text(S.of(context).drawerButtomSheetFCMToken),
-                subtitle: Text(S.of(context).clickToView),
+                title: Text(
+                    AppLocalizations.of(context)!.drawerButtomSheetFCMToken),
+                subtitle: Text(AppLocalizations.of(context)!.clickToView),
                 onTap: () {
                   showDialog(
                     context: context,
                     builder: (BuildContext context) => CustomDialog(
-                      title: S.of(context).drawerButtomSheetFCMToken,
+                      title: AppLocalizations.of(context)!
+                          .drawerButtomSheetFCMToken,
                       description: appState.fcmToken,
-                      buttonText: S.of(context).okButton,
+                      buttonText: AppLocalizations.of(context)!.okButton,
                     ),
                   );
                 },
               ),
               ListTile(
-                title: Text(S.of(context).drawerButtomSheetUserToken),
-                subtitle: Text(S.of(context).clickToView),
+                title: Text(
+                    AppLocalizations.of(context)!.drawerButtomSheetUserToken),
+                subtitle: Text(AppLocalizations.of(context)!.clickToView),
                 onTap: () {
                   showDialog(
                     context: context,
                     builder: (BuildContext context) => CustomDialog(
-                      title: S.of(context).drawerButtomSheetUserToken,
+                      title: AppLocalizations.of(context)!
+                          .drawerButtomSheetUserToken,
                       description: appState.userToken,
-                      buttonText: S.of(context).okButton,
+                      buttonText: AppLocalizations.of(context)!.okButton,
                     ),
                   );
                 },
               ),
               ListTile(
-                title: Text(S.of(context).drawerButtomSheetIdToken),
-                subtitle: Text(S.of(context).clickToView),
+                title: Text(
+                    AppLocalizations.of(context)!.drawerButtomSheetIdToken),
+                subtitle: Text(AppLocalizations.of(context)!.clickToView),
                 onTap: () {
                   showDialog(
                     context: context,
                     builder: (BuildContext context) => CustomDialog(
-                      title: S.of(context).drawerButtomSheetIdToken,
+                      title: AppLocalizations.of(context)!
+                          .drawerButtomSheetIdToken,
                       description: appState.idToken,
-                      buttonText: S.of(context).okButton,
+                      buttonText: AppLocalizations.of(context)!.okButton,
                     ),
                   );
                 },
               ),
               ListTile(
-                title: Text(S.of(context).drawerButtomSheetExpires),
+                title: Text(
+                    AppLocalizations.of(context)!.drawerButtomSheetExpires),
                 subtitle: Text(appState.expires == null
                     ? ''
                     : appState.expires!.toIso8601String()),