+++ lib/ui/pages/home/index.dart
@@ -1,47 +1,45 @@
+// ignore_for_file: unnecessary_new
+
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:first_app/models/appstate.dart';
 import 'package:first_app/models/locstate.dart';
 import 'package:first_app/ui/pages/login/index.dart';
-import 'package:first_app/generated/l10n.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:first_app/ui/pages/location/index.dart';
 import 'package:first_app/ui/pages/map/index.dart';
 import 'drawer.dart';
 
 class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     var appState = Provider.of<AppStateModel>(context);
     if (!appState.authenticated) {
-      return Scaffold(
+      return const Scaffold(
         body: LoginPage(),
       );
     }
 
-    // as documented in appstate.dart, we here set the defaultLokale
-    // from appState to apply loaded locale from sharedpreferences on
-    // startup.
-    Intl.defaultLocale = appState.locale;
-
     return Scaffold(
-        key: Key("HomePage_Scaffold"),
+        key: const Key("HomePage_Scaffold"),
         appBar: AppBar(
-          title: Text(S.of(context).appTitle),
+          title: Text(AppLocalizations.of(context)!.appTitle),
         ),
         backgroundColor: Theme.of(context).backgroundColor,
-        body: new ChangeNotifierProvider(
+        body: ChangeNotifierProvider(
           create: (_) => new LocStateModel(appState.mocks.getGeo()),
           child: Center(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
+              children: const <Widget>[
                 LocationStreamWidget(),
                 OverlayMapPage(),
               ],
             ),
           ),
         ),
-        drawer: HomePageDrawer());
+        drawer: const HomePageDrawer());
   }
 }