+++ lib/ui/pages/location/index.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:geolocator/geolocator.dart';
-import 'package:first_app/generated/l10n.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:first_app/models/locstate.dart';
 
 String latitudeLongitude(String lat, String long) => "Lat: $lat, Long: $long";
 
 class LocationStreamWidget extends StatelessWidget {
+  const LocationStreamWidget({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<bool>(
@@ -15,7 +17,7 @@ class LocationStreamWidget extends StatelessWidget {
           if (!snapshot.hasData) {
             return const Center(child: CircularProgressIndicator());
           }
-          return new Flexible(child: _buildListView(context));
+          return Flexible(child: _buildListView(context));
         });
   }
 
@@ -23,15 +25,15 @@ class LocationStreamWidget extends StatelessWidget {
     final List<Widget> listItems = <Widget>[
       ListTile(
         title: ElevatedButton(
-          key: Key('LocationPage_StartListeningButton'),
+          key: const Key('LocationPage_StartListeningButton'),
           style: ElevatedButton.styleFrom(
             elevation: 20.0,
             onPrimary: Theme.of(context).primaryColorLight,
             padding: const EdgeInsets.all(8.0),
           ),
           child: Text(Provider.of<LocStateModel>(context).isListening()
-              ? S.of(context).stopListening
-              : S.of(context).startListening),
+              ? AppLocalizations.of(context)!.stopListening
+              : AppLocalizations.of(context)!.startListening),
           onPressed: Provider.of<LocStateModel>(context).toggleListening,
         ),
       ),
@@ -39,7 +41,7 @@ class LocationStreamWidget extends StatelessWidget {
 
     listItems.addAll(Provider.of<LocStateModel>(context)
         .positions
-        .map((Position position) => PositionListItem(position))
+        .map((Position position) => PositionListItem(position: position))
         .toList());
 
     return ListView(
@@ -49,19 +51,18 @@ class LocationStreamWidget extends StatelessWidget {
 }
 
 class PositionListItem extends StatelessWidget {
-  const PositionListItem(this._position);
-
-  final Position _position;
+  final Position position;
+  const PositionListItem({Key? key, required this.position}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Container(
       padding:
           const EdgeInsets.only(left: 10.0, bottom: 2.0, top: 2.0, right: 10.0),
-      decoration: BoxDecoration(),
+      decoration: const BoxDecoration(),
       child: Card(
         elevation: 15.0,
-        margin: new EdgeInsets.symmetric(horizontal: 7.0, vertical: 7.0),
+        margin: const EdgeInsets.symmetric(horizontal: 7.0, vertical: 7.0),
         child: Padding(
           padding: const EdgeInsets.only(
             top: 2.0,
@@ -69,25 +70,25 @@ class PositionListItem extends StatelessWidget {
             left: 2.0,
           ),
           child: ListTile(
-            key: Key('LocationPage_LocationTile'),
+            key: const Key('LocationPage_LocationTile'),
             contentPadding:
-                EdgeInsets.symmetric(horizontal: 15.0, vertical: 3.0),
+                const EdgeInsets.symmetric(horizontal: 15.0, vertical: 3.0),
             title: Text(
               latitudeLongitude(
-                _position.latitude.toString(),
-                _position.longitude.toString(),
+                position.latitude.toString(),
+                position.longitude.toString(),
               ),
             ),
             subtitle: Row(
               children: <Widget>[
-                Icon(Icons.arrow_right),
+                const Icon(Icons.arrow_right),
                 Expanded(
                   child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       mainAxisAlignment: MainAxisAlignment.start,
                       children: <Widget>[
                         Text(Provider.of<LocStateModel>(context)
-                            .getAddressString(_position)),
+                            .getAddressString(position)),
                       ]),
                 ),
               ],