+++ test/location_test.dart
@@ -5,20 +5,20 @@ import 'package:provider/provider.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:first_app/ui/pages/location/index.dart';
 import 'package:first_app/mock/mock_geolocator.dart';
-import 'package:first_app/generated/l10n.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:first_app/ui/theme/style.dart';
 
 // Helper function to encapsulate code needed to instantiate the HomePage() widget
 dynamic initWidget(WidgetTester tester, LocStateModel locstate) {
   return tester.pumpWidget(
-    new MaterialApp(
-      onGenerateTitle: (context) => S.of(context).appTitle,
-      localizationsDelegates: [
-        S.delegate,
+    MaterialApp(
+      onGenerateTitle: (context) => AppLocalizations.of(context)!.appTitle,
+      localizationsDelegates: const [
+        AppLocalizations.delegate,
         GlobalMaterialLocalizations.delegate,
         GlobalWidgetsLocalizations.delegate,
       ],
-      supportedLocales: S.delegate.supportedLocales,
+      supportedLocales: AppLocalizations.supportedLocales,
       theme: appTheme,
       home: Scaffold(
         appBar: AppBar(),
@@ -27,7 +27,7 @@ dynamic initWidget(WidgetTester tester, LocStateModel locstate) {
           child: Center(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
+              children: const <Widget>[
                 LocationStreamWidget(),
               ],
             ),
@@ -46,7 +46,7 @@ void main() async {
     await tester.pump();
     expect(find.byType(LocationStreamWidget), findsOneWidget);
     // Wait for button to appear
-    await tester.pump(Duration(seconds: 1));
+    await tester.pump(const Duration(seconds: 1));
     expect(find.byType(ListTile), findsOneWidget);
     expect(
         find.descendant(
@@ -60,7 +60,7 @@ void main() async {
     await tester.tap(finder);
     expect(find.byType(ListView), findsOneWidget);
     // Wait for locations to appear from geolocator mock
-    await tester.pump(Duration(seconds: 20));
+    await tester.pump(const Duration(seconds: 20));
     // Even when specifying offstage widgets in find, only five are found,
     // although mock geolocator generates 10
     expect(